(implementation):create and set teams for todolist
