+++ app/lib/models/Team.dart
@@ -1,0 +1,8 @@
+/// Not yet fully implemented.
+class Team {
+  final String? name;
+
+  Team({
+    this.name,
+  });
+}
